// apps/admin/src/shared/api/generated/ja-korea/schemas/introductionUpdateRequest.ts
/**
 * Generated by orval v8.17.0 🍺
 * Do not edit manually.
 * JA Korea Homepage Admin API — JA Korea subset
 * Filtered for Homepage Admin JA Korea Orval codegen.
 * OpenAPI spec version: v1
 */

export interface IntroductionUpdateRequest {
  /**
     * @minLength 0
     * @maxLength 300
     */
  section1Title?: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  section1Subtitle?: string;
  /**
     * @minLength 0
     * @maxLength 300
     */
  section2TitleEn?: string;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  section2SubtitleKo?: string;
  /**
     * @minLength 0
     * @maxLength 300
     */
  content1Title?: string;
  /**
     * @minLength 0
     * @maxLength 10000
     */
  content1Description?: string;
  /**
     * @minLength 0
     * @maxLength 300
     */
  content2Title?: string;
  /**
     * @minLength 0
     * @maxLength 10000
     */
  content2Description?: string;
  /**
     * @minLength 0
     * @maxLength 300
     */
  globalVisionSubtitleEn?: string;
  /**
     * @minLength 0
     * @maxLength 10000
     */
  globalVisionTextKo?: string;
  /**
     * @minLength 0
     * @maxLength 300
     */
  globalMissionSubtitleEn?: string;
  /**
     * @minLength 0
     * @maxLength 10000
     */
  globalMissionTextKo?: string;
  version: number;
}